Refactored the code so that I substituted the uses of long for approximate reference addresses & PC addresses, & size of access with BigInteger to prevent overflow error when parsing the large hex numbers contained in the file and representing them with an appropriate form.

+++ src/analyzers/baseline_analyzer/InvalidAdditionToEmptyLoopStackException.java
@@ -1,13 +1,15 @@
 package analyzers.baseline_analyzer;
 
+import java.math.BigInteger;
+
 public class InvalidAdditionToEmptyLoopStackException extends Exception {
 
     private static final long serialVersionID = 7_718_828_512_143_293_558L;
     private long tripCount;
-    private long accessAddress;
+    private BigInteger accessAddress;
     private MemoryAccess accessType;
 
-    public InvalidAdditionToEmptyLoopStackException(long numTrips, long accessedAddr, MemoryAccess readOrWrite) {
+    public InvalidAdditionToEmptyLoopStackException(long numTrips, BigInteger accessedAddr, MemoryAccess readOrWrite) {
         super();
         this.tripCount = numTrips;
         this.accessAddress = accessedAddr;
@@ -15,21 +17,21 @@ public class InvalidAdditionToEmptyLoopStackException extends Exception {
 
     }
 
-    public InvalidAdditionToEmptyLoopStackException(String message, long numTrips, long accessedAddr, MemoryAccess readOrWrite) {
+    public InvalidAdditionToEmptyLoopStackException(String message, long numTrips, BigInteger accessedAddr, MemoryAccess readOrWrite) {
         super(message);
         this.tripCount = numTrips;
         this.accessAddress = accessedAddr;
         this.accessType = readOrWrite;
     }
 
-    public InvalidAdditionToEmptyLoopStackException(String message, Throwable cause, long numTrips, long accessedAddr, MemoryAccess readOrWrite) {
+    public InvalidAdditionToEmptyLoopStackException(String message, Throwable cause, long numTrips, BigInteger accessedAddr, MemoryAccess readOrWrite) {
         super(message, cause);
         this.tripCount = numTrips;
         this.accessAddress = accessedAddr;
         this.accessType = readOrWrite;
     }
 
-    public InvalidAdditionToEmptyLoopStackException(Throwable cause, long numTrips, long accessedAddr, MemoryAccess readOrWrite) {
+    public InvalidAdditionToEmptyLoopStackException(Throwable cause, long numTrips, BigInteger accessedAddr, MemoryAccess readOrWrite) {
         super(cause);
         this.tripCount = numTrips;
         this.accessAddress = accessedAddr;
@@ -40,7 +42,7 @@ public class InvalidAdditionToEmptyLoopStackException extends Exception {
         return tripCount;
     }
 
-    public long getAccessAddress() {
+    public BigInteger getAccessAddress() {
         return accessAddress;
     }
